#4 refactor and tests
