+++ src/test/java/com/community/bitcoinwallet/service/H2WalletServiceTest.java
@@ -1,11 +1,9 @@
-package com.community.bitcoinwallet.repository;
+package com.community.bitcoinwallet.service;
 
 import com.community.bitcoinwallet.BitcoinWalletApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
-
 @SpringBootTest(classes = {BitcoinWalletApplication.class},
     properties = {"spring.profiles.active=h2", "bitcoin-wallet.balance.async-balance=false"})
-public class H2WalletRepositoryTest extends GenericRepositoryTest {
-
+public class H2WalletServiceTest extends WalletServiceTest {
 }
