+++ src/test/java/com/community/bitcoinwallet/service/InMemoryWalletServiceTest.java
@@ -1,10 +1,9 @@
-package com.community.bitcoinwallet.repository;
+package com.community.bitcoinwallet.service;
 
 import com.community.bitcoinwallet.BitcoinWalletApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
-
 @SpringBootTest(classes = {BitcoinWalletApplication.class},
     properties = {"spring.profiles.active=in-memory", "bitcoin-wallet.balance.async-balance=false"})
-public class InMemoryWalletRepositoryTest extends GenericRepositoryTest {
+public class InMemoryWalletServiceTest extends WalletServiceTest {
 }
