+++ src/test/java/com/community/bitcoinwallet/service/WalletServiceTest.java
@@ -1,6 +1,5 @@
 package com.community.bitcoinwallet.service;
 
-import com.community.bitcoinwallet.BitcoinWalletApplication;
 import com.community.bitcoinwallet.model.WalletEntry;
 import com.community.bitcoinwallet.repository.WalletRepository;
 import com.community.bitcoinwallet.util.DateAndAmountUtils;
@@ -9,7 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,10 +17,8 @@ import java.util.Arrays;
 import java.util.Collections;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = {BitcoinWalletApplication.class},
-    properties = {"spring.profiles.active=h2","bitcoin-wallet.balance.async-balance=false"})
 @Transactional
-public class WalletServiceTest {
+public abstract class WalletServiceTest {
 
     private static final int COUNT_ENTRIES = 6;
     private static final int QUARTER_OF_HOUR_SECONDS = 15 * 60;
